+++ lib/types/atom.d.ts
@@ -1,0 +1,14 @@
+import { TextEditor } from 'atom'
+
+// Missing Atom API
+declare module 'atom' {
+  interface CompositeDisposable {
+    disposed: boolean
+  }
+  interface Pane {
+    getPendingItem(): TextEditor
+  }
+  interface Notification {
+    getOptions(): { detail: string }
+  }
+}
